途中

+++ src/graph/graph.hpp
@@ -113,7 +113,7 @@ public:
      */
     constexpr Vec<T> depths(int R = 0) const
     {
-        Vec<int> Ans(V(), 0);
+        Vec<i64> Ans(V(), 0);
         Fix([&](auto self, int u, int p) -> void {
             for (const auto& [v, c] : m_lists[u] | ToC) {
                 if (v == p) { continue; }
